Add files via upload
It's here. This is a test version of the new Mrs Jennings update. It's certainly not the final version and I may even scrap it and completely redo it (again) if people don't like it. I need you guys to test it out and report bugs. As for the changes, most of them are internal but there are a couple noteworthy things to add. 

Mrs Jennings now has staff meetings on Wednesday in the Afternoon. After a certain point you can interrupt the staff meeting to talk with Mrs Jennings.

Mrs Jennings Handjob and Blowjobs are repeatable in class after the Toilet scene in school. 

You can now choose which scene you want to watch when visiting Mrs Jennings quest after you view them all. 

Fixed up the quest so the tutoring scene didn't seem so jumping. Sorry if it seems out of character but dialogue is not my forte

Reactivated noon scenes. You can do them by talking to Mrs Jennings during recess.

Remember this is going to be improved upon so take it easy alright?